Split PoetClientFunctionalTests into different tests
